refactor(streamrun): extract the replication engine into internal/streamrun
Step 1 of decoupling the web console into its own binary: move the
self-contained streaming engine out of cmd/bintrail (package main) into a
reusable internal/streamrun package, so a future bintrail-console binary can
run supervised streams without importing the cmd layer. Core bintrail (what
dbtrail runs) is unaffected.

Pure move, zero behavior change:
- internal/streamrun: streamConfig→Config, streamHooks→Hooks, streamOne→One,
  plus streamState/loadStreamState/saveCheckpoint, buildTLSConfig,
  resolveStart(+AutoDiscover), gap detection, streamLoop, normalizeGTIDSet→
  NormalizeGTIDSet, startMetricsServer→StartMetricsServer — moved verbatim.
- streamOne's host-supplied helpers (source preflight/validation, identity
  resolution, schema snapshot/filters, DDL recording, source-DSN parse, JSON
  summary) are INJECTED via a Config.Deps seam rather than dragged into
  streamrun — they stay in cmd/bintrail (shared with `index`); streamDeps()
  wires them. Relocating them to internal/ is a later step (PR-B).
- cmd/bintrail keeps the cobra layer: streamCmd/flags, parseSourceDSN,
  streamConfigFromFlags (now returns streamrun.Config), runStream. up.go and
  monitor.go call streamrun.{One,Config,Hooks,StartMetricsServer,
  NormalizeGTIDSet}.
- Engine tests moved with the engine (unit + integration); cmd-layer tests
  (parseSourceDSN, streamCmd wiring, the flag→config snapshot, One's
  validation) stay in cmd/bintrail.

Verified: go build/vet (unit + integration) clean; full unit suite green;
streamrun integration tests (incl. live replication) and the monitor/
supervisor integration tests (which exercise the Deps seam end-to-end) pass
against MySQL.
