+++ cmd/bintrail/up.go
@@ -19,6 +19,7 @@ import (
 	"github.com/dbtrail/bintrail/internal/config"
 	"github.com/dbtrail/bintrail/internal/console"
 	"github.com/dbtrail/bintrail/internal/indexer"
+	"github.com/dbtrail/bintrail/internal/streamrun"
 )
 
 var upCmd = &cobra.Command{
@@ -287,7 +288,7 @@ func runUpConsoleOnly(cmd *cobra.Command) error {
 	// (and would fight over the bind). Synchronous bind: fails fast, like
 	// the console bind.
 	if upMetricsAddr != "" {
-		stopMetrics, err := startMetricsServer(upMetricsAddr)
+		stopMetrics, err := streamrun.StartMetricsServer(upMetricsAddr)
 		if err != nil {
 			return err
 		}
@@ -414,7 +415,7 @@ func runUpStreamWithConsole(cmd *cobra.Command, args []string) error {
 	// double-bind the same address inside streamOne. Synchronous bind:
 	// fails fast, like the console bind below.
 	if upMetricsAddr != "" {
-		stopMetrics, err := startMetricsServer(upMetricsAddr)
+		stopMetrics, err := streamrun.StartMetricsServer(upMetricsAddr)
 		if err != nil {
 			return err
 		}
