+++ internal/streamrun/streamrun.go
@@ -1,0 +1,1253 @@
+package streamrun
+
+import (
+	"context"
+	"crypto/tls"
+	"crypto/x509"
+	"database/sql"
+	"errors"
+	"fmt"
+	"log/slog"
+	"net"
+	"net/http"
+	"os"
+	"strings"
+	"time"
+
+	gomysql "github.com/go-mysql-org/go-mysql/mysql"
+	"github.com/go-mysql-org/go-mysql/replication"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+
+	"github.com/dbtrail/bintrail/internal/cliutil"
+	"github.com/dbtrail/bintrail/internal/config"
+	"github.com/dbtrail/bintrail/internal/indexer"
+	"github.com/dbtrail/bintrail/internal/metadata"
+	"github.com/dbtrail/bintrail/internal/observe"
+	"github.com/dbtrail/bintrail/internal/parser"
+	"github.com/dbtrail/bintrail/internal/serverid"
+)
+
+// ─── streamState ───────────────────────────────────────────────────────────────────
+
+// streamState holds the current replication position and counters used for
+// checkpointing. It is persisted to stream_state after each checkpoint interval.
+type streamState struct {
+	mode          string // "position" or "gtid"
+	binlogFile    string
+	binlogPos     uint64
+	gtidSet       string // serialized GTID set (GTID mode only)
+	eventsIndexed int64
+	lastEventTime sql.NullTime
+	serverID      uint32
+	bintrailID    string // resolved server identity (empty = unknown, stored as NULL)
+
+	// accGTID is the in-memory accumulated GTID set (GTID mode only).
+	// It is serialized to gtidSet on checkpoint.
+	accGTID *gomysql.MysqlGTIDSet
+}
+
+// loadStreamState loads the saved stream_state row, returning nil if no row exists.
+func loadStreamState(db *sql.DB) (*streamState, error) {
+	var s streamState
+	var gtidSet, bintrailID sql.NullString
+	err := db.QueryRow(`
+		SELECT mode, binlog_file, binlog_position, gtid_set,
+		       events_indexed, last_event_time, server_id, bintrail_id
+		FROM stream_state WHERE id = 1`).Scan(
+		&s.mode, &s.binlogFile, &s.binlogPos, &gtidSet,
+		&s.eventsIndexed, &s.lastEventTime, &s.serverID, &bintrailID)
+	if errors.Is(err, sql.ErrNoRows) {
+		return nil, nil
+	}
+	if err != nil {
+		return nil, fmt.Errorf("query stream_state: %w", err)
+	}
+	if gtidSet.Valid {
+		s.gtidSet = gtidSet.String
+	}
+	if bintrailID.Valid {
+		s.bintrailID = bintrailID.String
+	}
+	return &s, nil
+}
+
+// saveCheckpoint persists the current stream state to the stream_state table.
+func saveCheckpoint(db *sql.DB, state *streamState) error {
+	var gtidSet any
+	if state.gtidSet != "" {
+		gtidSet = state.gtidSet
+	}
+	var lastEventTime any
+	if state.lastEventTime.Valid {
+		lastEventTime = state.lastEventTime.Time
+	}
+	var bintrailIDArg any
+	if state.bintrailID != "" {
+		bintrailIDArg = state.bintrailID
+	}
+	_, err := db.Exec(`
+		INSERT INTO stream_state
+		    (id, mode, binlog_file, binlog_position, gtid_set,
+		     events_indexed, last_event_time, last_checkpoint, server_id, bintrail_id)
+		VALUES (1, ?, ?, ?, ?, ?, ?, UTC_TIMESTAMP(), ?, ?)
+		ON DUPLICATE KEY UPDATE
+		    binlog_file     = VALUES(binlog_file),
+		    binlog_position = VALUES(binlog_position),
+		    gtid_set        = VALUES(gtid_set),
+		    events_indexed  = VALUES(events_indexed),
+		    last_event_time = VALUES(last_event_time),
+		    last_checkpoint = UTC_TIMESTAMP(),
+		    server_id       = VALUES(server_id),
+		    bintrail_id     = VALUES(bintrail_id)`,
+		state.mode, state.binlogFile, state.binlogPos, gtidSet,
+		state.eventsIndexed, lastEventTime, state.serverID, bintrailIDArg)
+	return err
+}
+
+// ─── TLS configuration ───────────────────────────────────────────────────────────────
+
+// buildTLSConfig returns a *tls.Config for the given ssl-mode, or nil for
+// "disabled". serverName is the target host (used only for verify-identity).
+func buildTLSConfig(mode, ca, cert, key, serverName string) (*tls.Config, error) {
+	if mode == "disabled" {
+		return nil, nil
+	}
+	switch mode {
+	case "preferred", "required", "verify-ca", "verify-identity":
+	default:
+		return nil, fmt.Errorf("invalid --ssl-mode %q: must be one of disabled, preferred, required, verify-ca, verify-identity", mode)
+	}
+	if (cert == "") != (key == "") {
+		return nil, fmt.Errorf("--ssl-cert and --ssl-key must both be specified together")
+	}
+
+	cfg := &tls.Config{}
+
+	// Load CA pool (optional — system CAs used when empty).
+	var caPool *x509.CertPool
+	if ca != "" {
+		pem, err := os.ReadFile(ca)
+		if err != nil {
+			return nil, fmt.Errorf("read --ssl-ca %q: %w", ca, err)
+		}
+		caPool = x509.NewCertPool()
+		if !caPool.AppendCertsFromPEM(pem) {
+			return nil, fmt.Errorf("--ssl-ca %q: no valid certificates found", ca)
+		}
+		cfg.RootCAs = caPool
+	}
+
+	// Load client certificate for mutual TLS.
+	if cert != "" {
+		kp, err := tls.LoadX509KeyPair(cert, key)
+		if err != nil {
+			return nil, fmt.Errorf("load --ssl-cert/--ssl-key: %w", err)
+		}
+		cfg.Certificates = []tls.Certificate{kp}
+	}
+
+	switch mode {
+	case "preferred", "required":
+		// Encrypt the connection but skip server certificate verification.
+		cfg.InsecureSkipVerify = true //nolint:gosec // intentional for these modes
+	case "verify-ca":
+		// Verify the certificate chain against the CA pool but not the hostname.
+		cfg.InsecureSkipVerify = true //nolint:gosec // hostname check done via VerifyConnection
+		cfg.VerifyConnection = func(cs tls.ConnectionState) error {
+			if len(cs.PeerCertificates) == 0 {
+				return errors.New("server presented no certificate")
+			}
+			opts := x509.VerifyOptions{
+				Roots:         caPool, // nil → system CAs
+				Intermediates: x509.NewCertPool(),
+			}
+			for _, c := range cs.PeerCertificates[1:] {
+				opts.Intermediates.AddCert(c)
+			}
+			_, err := cs.PeerCertificates[0].Verify(opts)
+			return err
+		}
+	case "verify-identity":
+		// Full TLS verification: certificate chain + hostname.
+		cfg.ServerName = serverName
+	}
+
+	return cfg, nil
+}
+
+// ─── Start position resolution ───────────────────────────────────────────────
+
+// NormalizeGTIDSet zero-pads each UUID in a GTID set to the standard
+// 8-4-4-4-12 format. Some MySQL-compatible services (e.g. Amazon RDS) return
+// GTIDs with leading zeros stripped from UUID segments, producing UUIDs shorter
+// than 36 characters (e.g. "5512139-1432-11f1-8d8d-0693b428a89b" instead of
+// "05512139-1432-11f1-8d8d-0693b428a89b"). The go-mysql library requires
+// standard-length UUIDs, so this function normalizes before parsing.
+func NormalizeGTIDSet(s string) string {
+	// Expected segment lengths in a UUID: 8-4-4-4-12.
+	segLens := [5]int{8, 4, 4, 4, 12}
+
+	// A GTID set is comma-separated entries like "uuid:intervals,uuid:intervals".
+	entries := strings.Split(s, ",")
+	for i, entry := range entries {
+		entry = strings.TrimSpace(entry)
+		if entry == "" {
+			continue
+		}
+		// Split "uuid:intervals" at the first colon.
+		colon := strings.IndexByte(entry, ':')
+		if colon < 0 {
+			continue // malformed, let the parser handle it
+		}
+		uuid := entry[:colon]
+		rest := entry[colon:] // includes the ":"
+
+		parts := strings.Split(uuid, "-")
+		if len(parts) != 5 {
+			continue // not a UUID shape, let the parser handle it
+		}
+
+		changed := false
+		for j, seg := range parts {
+			if len(seg) < segLens[j] {
+				parts[j] = strings.Repeat("0", segLens[j]-len(seg)) + seg
+				changed = true
+			}
+		}
+		if changed {
+			entries[i] = strings.Join(parts, "-") + rest
+		} else {
+			entries[i] = entry
+		}
+	}
+	return strings.Join(entries, ",")
+}
+
+// resolveStart determines the start position for replication. It returns the
+// mode ("position" or "gtid"), file, GTID string, pos, and an optional
+// pre-parsed MysqlGTIDSet (non-nil only in GTID mode).
+func resolveStart(
+	startFile, startGTID string, startPos uint32,
+	saved *streamState,
+) (mode, file, gtidStr string, pos uint32, accGTID *gomysql.MysqlGTIDSet, err error) {
+	// Saved checkpoint takes priority — makes re-running the same command
+	// idempotent (the user doesn't need to remove --start-file to resume).
+	// Exception: if the user explicitly requests a *different* mode than the
+	// saved checkpoint (e.g. saved=position but --start-gtid given, or
+	// saved=gtid but --start-file given), honor the flag to allow seamless
+	// mode switching without deleting stream_state.
+	if saved != nil {
+		if startFile != "" && startGTID != "" {
+			return "", "", "", 0, nil, fmt.Errorf("--start-file and --start-gtid are mutually exclusive")
+		}
+
+		// Detect mode switch: user explicitly requests a different mode.
+		switchToGTID := saved.mode == "position" && startGTID != "" && startFile == ""
+		switchToPosition := saved.mode == "gtid" && startFile != "" && startGTID == ""
+
+		if switchToGTID {
+			slog.Warn("switching from position mode to GTID mode", "old_file", saved.binlogFile, "old_pos", saved.binlogPos)
+			startGTID = NormalizeGTIDSet(startGTID)
+			gs, parseErr := gomysql.ParseMysqlGTIDSet(startGTID)
+			if parseErr != nil {
+				return "", "", "", 0, nil, fmt.Errorf("invalid --start-gtid: %w", parseErr)
+			}
+			return "gtid", "", startGTID, 0, gs.(*gomysql.MysqlGTIDSet), nil
+		}
+		if switchToPosition {
+			slog.Warn("switching from GTID mode to position mode", "old_gtid_set", saved.gtidSet)
+			return "position", startFile, "", startPos, nil, nil
+		}
+
+		// Same mode or no flags — resume from saved state.
+		if startFile != "" || startGTID != "" {
+			slog.Warn("checkpoint exists; ignoring --start-file/--start-gtid and resuming from saved state")
+		}
+		if saved.mode == "gtid" {
+			normalized := NormalizeGTIDSet(saved.gtidSet)
+			slog.Info("resuming from GTID set", "gtid_set", normalized)
+			gs, parseErr := gomysql.ParseMysqlGTIDSet(normalized)
+			if parseErr != nil {
+				return "", "", "", 0, nil, fmt.Errorf("invalid saved gtid_set %q: %w", saved.gtidSet, parseErr)
+			}
+			return "gtid", "", normalized, 0, gs.(*gomysql.MysqlGTIDSet), nil
+		}
+		slog.Info("resuming from position", "file", saved.binlogFile, "pos", saved.binlogPos)
+		return "position", saved.binlogFile, "", uint32(saved.binlogPos), nil, nil
+	}
+
+	// No checkpoint — use flags for initial start position (first run).
+	if startFile != "" && startGTID != "" {
+		return "", "", "", 0, nil, fmt.Errorf("--start-file and --start-gtid are mutually exclusive")
+	}
+	if startGTID != "" {
+		startGTID = NormalizeGTIDSet(startGTID)
+		gs, parseErr := gomysql.ParseMysqlGTIDSet(startGTID)
+		if parseErr != nil {
+			return "", "", "", 0, nil, fmt.Errorf("invalid --start-gtid: %w", parseErr)
+		}
+		return "gtid", "", startGTID, 0, gs.(*gomysql.MysqlGTIDSet), nil
+	}
+	if startFile != "" {
+		return "position", startFile, "", startPos, nil, nil
+	}
+
+	return "", "", "", 0, nil, fmt.Errorf(
+		"no start position specified and no saved stream state found; " +
+			"provide --start-file or --start-gtid to begin streaming")
+}
+
+// resolveStartWithAutoDiscover wraps resolveStart with an auto-discover
+// fallback for the first-run, no-flags case. When neither --start-file nor
+// --start-gtid is set AND no checkpoint exists yet, the provided autoDiscover
+// callback is invoked to query the source's current binlog position (matching
+// the behavior of `bintrail agent` BYOS mode). The callback is typically
+// config.CurrentBinlogPosition(sourceDB).
+//
+// All other paths (saved checkpoint, explicit flags, mutually-exclusive flags,
+// invalid GTID, etc.) delegate to resolveStart unchanged. The wrapper exists
+// so the discovery side-effect can be unit-tested without a real *sql.DB —
+// callers pass a stub function.
+func resolveStartWithAutoDiscover(
+	startFile, startGTID string, startPos uint32,
+	saved *streamState,
+	autoDiscover func() (string, uint32, error),
+) (mode, file, gtidStr string, pos uint32, accGTID *gomysql.MysqlGTIDSet, err error) {
+	mode, file, gtidStr, pos, accGTID, err = resolveStart(startFile, startGTID, startPos, saved)
+	if err == nil {
+		return
+	}
+	// Only fall back to auto-discover when the failure is the specific
+	// "no start position" case (saved == nil and no flags). Other errors
+	// (mutually-exclusive flags, invalid GTID, corrupt saved state) must
+	// propagate so the operator sees the real problem.
+	if autoDiscover == nil || saved != nil || startFile != "" || startGTID != "" {
+		return
+	}
+	af, ap, dErr := autoDiscover()
+	if dErr != nil {
+		return "", "", "", 0, nil, fmt.Errorf("auto-discover binlog position: %w", dErr)
+	}
+	return "position", af, "", ap, nil, nil
+}
+
+// ─── Gap detection ──────────────────────────────────────────────────────────────
+
+// gapResult describes a binlog gap between the saved checkpoint and the source.
+type gapResult struct {
+	HasGap   bool   // true if a gap exists between checkpoint and source
+	Fillable bool   // true if the gap can be filled (binlogs still available)
+	Message  string // human-readable description of the gap
+
+	// For unfillable gaps in position mode: the earliest available binlog.
+	EarliestFile string
+	EarliestPos  uint32
+
+	// For unfillable gaps in GTID mode: the purged GTID set to use as base.
+	PurgedGTIDSet string
+}
+
+// detectPositionGap queries the source MySQL for available binary logs and
+// checks whether the checkpoint file still exists. Returns a gapResult.
+// timeout caps how long SHOW BINARY LOGS may take — RDS instances with many
+// binlog files can take >10s, so callers should pass a generous value.
+func detectPositionGap(sourceDB *sql.DB, checkpointFile string, checkpointPos uint32, timeout time.Duration) (*gapResult, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	rows, err := sourceDB.QueryContext(ctx, "SHOW BINARY LOGS")
+	if err != nil {
+		return nil, fmt.Errorf("SHOW BINARY LOGS: %w", err)
+	}
+	defer rows.Close()
+
+	// SHOW BINARY LOGS returns varying columns across MySQL versions;
+	// we only need the first two (Log_name, File_size).
+	cols, colErr := rows.Columns()
+	if colErr != nil {
+		return nil, fmt.Errorf("SHOW BINARY LOGS columns: %w", colErr)
+	}
+	if len(cols) < 2 {
+		return nil, fmt.Errorf("SHOW BINARY LOGS returned %d columns, expected at least 2", len(cols))
+	}
+
+	type binlogEntry struct {
+		name string
+		size int64
+	}
+	var logs []binlogEntry
+	for rows.Next() {
+		var name string
+		var size int64
+		vals := make([]any, len(cols))
+		vals[0] = &name
+		vals[1] = &size
+		for i := 2; i < len(cols); i++ {
+			vals[i] = new(sql.RawBytes)
+		}
+		if err := rows.Scan(vals...); err != nil {
+			return nil, fmt.Errorf("scan SHOW BINARY LOGS: %w", err)
+		}
+		logs = append(logs, binlogEntry{name, size})
+	}
+	if err := rows.Err(); err != nil {
+		return nil, fmt.Errorf("iterate SHOW BINARY LOGS: %w", err)
+	}
+	if len(logs) == 0 {
+		return nil, fmt.Errorf("SHOW BINARY LOGS returned no results")
+	}
+
+	// Check if the checkpoint file exists in the list.
+	for _, l := range logs {
+		if l.name == checkpointFile {
+			// File exists but checkpoint position exceeds file size — the file
+			// may have been regenerated (e.g., RESET MASTER). Treat as unfillable.
+			if int64(checkpointPos) > l.size {
+				return &gapResult{
+					HasGap:       true,
+					Fillable:     false,
+					EarliestFile: logs[0].name,
+					EarliestPos:  4,
+					Message: fmt.Sprintf(
+						"binlog gap: file %s exists but checkpoint position %d exceeds file size %d; "+
+							"file may have been regenerated after RESET MASTER",
+						checkpointFile, checkpointPos, l.size),
+				}, nil
+			}
+			// File still exists and position is valid — gap is fillable.
+			currentFile := logs[len(logs)-1].name
+			if checkpointFile == currentFile {
+				return &gapResult{HasGap: false}, nil
+			}
+			return &gapResult{
+				HasGap:   true,
+				Fillable: true,
+				Message: fmt.Sprintf(
+					"gap detected: checkpoint is at %s:%d, source is at %s; replaying missed events",
+					checkpointFile, checkpointPos, currentFile),
+			}, nil
+		}
+	}
+
+	// File not found — binlogs have been purged.
+	earliest := logs[0]
+	return &gapResult{
+		HasGap:       true,
+		Fillable:     false,
+		EarliestFile: earliest.name,
+		EarliestPos:  4, // first 4 bytes are the magic number header, not event data
+		Message: fmt.Sprintf(
+			"binlog gap detected but CANNOT be filled: required file %s has been purged; "+
+				"earliest available binlog is %s; events between these positions are permanently lost",
+			checkpointFile, earliest.name),
+	}, nil
+}
+
+// gtidSetsEqual parses two GTID set strings and compares them structurally,
+// avoiding false mismatches from formatting differences (UUID case, ordering).
+// Returns false (not an error) if either string cannot be parsed as a valid
+// GTID set — the caller should treat this as "not equal" and proceed with
+// gap detection.
+func gtidSetsEqual(a, b string) bool {
+	if a == "" && b == "" {
+		return true
+	}
+	ga, err := gomysql.ParseMysqlGTIDSet(NormalizeGTIDSet(a))
+	if err != nil {
+		slog.Debug("gtidSetsEqual: failed to parse first GTID set", "gtid_set", a, "error", err)
+		return false
+	}
+	gb, err := gomysql.ParseMysqlGTIDSet(NormalizeGTIDSet(b))
+	if err != nil {
+		slog.Debug("gtidSetsEqual: failed to parse second GTID set", "gtid_set", b, "error", err)
+		return false
+	}
+	return ga.Equal(gb)
+}
+
+// detectGTIDGap queries the source MySQL for @@gtid_purged and @@gtid_executed,
+// then checks whether the checkpoint GTID set requires any purged GTIDs.
+// timeout caps how long the GTID system-variable queries may take.
+func detectGTIDGap(sourceDB *sql.DB, checkpointGTID string, timeout time.Duration) (*gapResult, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	checkpointGTID = NormalizeGTIDSet(strings.TrimSpace(checkpointGTID))
+
+	var purgedStr, executedStr string
+	if err := sourceDB.QueryRowContext(ctx, "SELECT @@gtid_purged").Scan(&purgedStr); err != nil {
+		return nil, fmt.Errorf("query @@gtid_purged: %w", err)
+	}
+	if err := sourceDB.QueryRowContext(ctx, "SELECT @@gtid_executed").Scan(&executedStr); err != nil {
+		return nil, fmt.Errorf("query @@gtid_executed: %w", err)
+	}
+
+	purgedStr = NormalizeGTIDSet(strings.TrimSpace(purgedStr))
+	executedStr = NormalizeGTIDSet(strings.TrimSpace(executedStr))
+
+	// If purged is empty, all GTIDs are still available.
+	if purgedStr == "" {
+		if gtidSetsEqual(checkpointGTID, executedStr) {
+			return &gapResult{HasGap: false}, nil
+		}
+		return &gapResult{
+			HasGap:   true,
+			Fillable: true,
+			Message:  "gap detected: checkpoint GTID set is behind source @@gtid_executed; replaying missed events",
+		}, nil
+	}
+
+	if checkpointGTID == "" {
+		return nil, fmt.Errorf("checkpoint GTID set is empty; cannot perform gap detection")
+	}
+
+	// Parse both GTID sets.
+	checkpoint, err := gomysql.ParseMysqlGTIDSet(checkpointGTID)
+	if err != nil {
+		return nil, fmt.Errorf("parse checkpoint GTID set: %w", err)
+	}
+	purged, err := gomysql.ParseMysqlGTIDSet(purgedStr)
+	if err != nil {
+		return nil, fmt.Errorf("parse @@gtid_purged: %w", err)
+	}
+
+	cpSet, ok := checkpoint.(*gomysql.MysqlGTIDSet)
+	if !ok {
+		return nil, fmt.Errorf("unexpected GTID set type %T from checkpoint parse", checkpoint)
+	}
+	purgedSet, ok := purged.(*gomysql.MysqlGTIDSet)
+	if !ok {
+		return nil, fmt.Errorf("unexpected GTID set type %T from purged parse", purged)
+	}
+
+	// Check if the stream would need any purged GTIDs on resume. MySQL sends
+	// all GTIDs NOT in the checkpoint set, so we must check two directions:
+	//
+	// 1. Forward: for each UUID in the checkpoint, does the purged set contain
+	//    any transaction IDs that the checkpoint doesn't have? If so, MySQL
+	//    would need to send those GTIDs but they've been purged.
+	// 2. Reverse: are there UUIDs in the purged set that the checkpoint has
+	//    never seen? MySQL would try to send all of that UUID's GTIDs, but
+	//    some have been purged.
+	needsPurged := false
+
+	// Direction 1: for each UUID in the purged set, check if the checkpoint
+	// fully covers all purged intervals. If the checkpoint doesn't contain all
+	// purged GTIDs for a UUID, MySQL would need to send them but they're gone.
+	// This correctly handles non-contiguous intervals (e.g., checkpoint has
+	// uuid:1-100,200-300 but purged has uuid:1-150 — GTIDs 101-150 are lost).
+	//
+	// Direction 2 (inline): if a purged UUID is absent from the checkpoint
+	// entirely, MySQL would try to send all of that UUID's GTIDs.
+	for uuid, purgedIntervals := range purgedSet.Sets {
+		cpIntervals, exists := cpSet.Sets[uuid]
+		if !exists {
+			// Purged UUID the checkpoint has never seen.
+			needsPurged = true
+			break
+		}
+		if len(purgedIntervals.Intervals) == 0 {
+			continue
+		}
+		if len(cpIntervals.Intervals) == 0 {
+			needsPurged = true
+			break
+		}
+		// Check if the checkpoint's intervals fully contain the purged intervals.
+		// IntervalSlice.Contain(sub) returns true if sub is a subset of s.
+		if !cpIntervals.Intervals.Contain(purgedIntervals.Intervals) {
+			needsPurged = true
+			break
+		}
+	}
+
+	if needsPurged {
+		return &gapResult{
+			HasGap:        true,
+			Fillable:      false,
+			PurgedGTIDSet: purgedStr,
+			Message: fmt.Sprintf(
+				"GTID gap detected but CANNOT be filled: required GTIDs have been purged from the source; "+
+					"purged set: %s; events in the purged range are permanently lost",
+				purgedStr),
+		}, nil
+	}
+
+	// Checkpoint is ahead of (or equal to) the purged set — gap is fillable.
+	if gtidSetsEqual(checkpointGTID, executedStr) {
+		return &gapResult{HasGap: false}, nil
+	}
+	return &gapResult{
+		HasGap:   true,
+		Fillable: true,
+		Message:  "gap detected: checkpoint GTID set is behind source @@gtid_executed; replaying missed events",
+	}, nil
+}
+
+// ─── Stream loop ────────────────────────────────────────────────────────────────
+
+// streamLoop consumes parser events, flushes batches to MySQL, and writes
+// checkpoints to stream_state at the given interval.
+//
+// DDL events flush the current batch and are never inserted. The DDL
+// HANDLING (auto-snapshot + resolver swap + schema_changes record) is NOT
+// done here on purpose: this loop runs behind the parser through a buffered
+// channel, so by the time a DDL reaches it the parser has already decoded —
+// and, for a new table, skipped — the rows that followed the DDL in the
+// binlog. That work lives in the parser's synchronous DDL hook
+// (StreamParser.SetSyncDDLHook, #396).
+func streamLoop(
+	ctx context.Context,
+	events <-chan parser.Event,
+	idx *indexer.Indexer,
+	db *sql.DB,
+	checkpointInterval time.Duration,
+	state *streamState,
+	m *observe.StreamMetrics,
+	hooks *Hooks,
+) error {
+	batch := make([]parser.Event, 0, idx.BatchSize())
+	ticker := time.NewTicker(checkpointInterval)
+	defer ticker.Stop()
+
+	flush := func() error {
+		if len(batch) == 0 {
+			return nil
+		}
+		m.BatchSize.Observe(float64(len(batch)))
+		n, err := idx.InsertBatch(batch)
+		state.eventsIndexed += n
+		m.EventsIndexed.Add(float64(n))
+		m.BatchFlushes.Inc()
+		batch = batch[:0]
+		if err != nil {
+			m.Errors.WithLabelValues("batch_flush").Inc()
+			return err
+		}
+		if n > 0 && hooks != nil && hooks.OnIndexed != nil {
+			hooks.OnIndexed(n)
+		}
+		return nil
+	}
+
+	checkpoint := func() {
+		if err := flush(); err != nil {
+			slog.Warn("batch flush failed", "error", err)
+			return
+		}
+		if err := saveCheckpoint(db, state); err != nil {
+			slog.Warn("saveCheckpoint failed", "error", err)
+			m.Errors.WithLabelValues("checkpoint").Inc()
+		} else {
+			m.CheckpointSaves.Inc()
+			if hooks != nil && hooks.OnCheckpoint != nil {
+				hooks.OnCheckpoint()
+			}
+			slog.Info("checkpoint saved",
+				"file", state.binlogFile,
+				"pos", state.binlogPos,
+				"events_indexed", state.eventsIndexed)
+		}
+	}
+
+	for {
+		select {
+		case <-ctx.Done():
+			checkpoint()
+			return nil
+
+		case <-ticker.C:
+			checkpoint()
+
+		case ev, ok := <-events:
+			if !ok {
+				checkpoint()
+				return nil
+			}
+			// Update position tracking from each event.
+			if ev.BinlogFile != "" {
+				state.binlogFile = ev.BinlogFile
+			}
+			state.binlogPos = ev.EndPos
+			if ev.GTID != "" && state.accGTID != nil {
+				if err := state.accGTID.Update(ev.GTID); err != nil {
+					slog.Warn("failed to update GTID set", "gtid", ev.GTID, "error", err)
+					m.Errors.WithLabelValues("gtid_update").Inc()
+				} else {
+					state.gtidSet = state.accGTID.String()
+				}
+			}
+			if !ev.Timestamp.IsZero() {
+				state.lastEventTime = sql.NullTime{Time: ev.Timestamp, Valid: true}
+			}
+
+			// GTID-only events: position/GTID already tracked above, skip insertion.
+			if ev.EventType == parser.EventGTID {
+				continue
+			}
+
+			// DDL events: flush batch, skip insertion (handled by the
+			// parser's synchronous hook — see the function comment).
+			if ev.EventType == parser.EventDDL {
+				if err := flush(); err != nil {
+					return err
+				}
+				continue
+			}
+
+			m.EventsReceived.Inc()
+			if !ev.Timestamp.IsZero() {
+				ts := float64(ev.Timestamp.Unix())
+				m.LastEventTimestamp.Set(ts)
+				m.ReplicationLag.Set(float64(time.Now().Unix()) - ts)
+			}
+			slog.Debug("event received",
+				"schema", ev.Schema,
+				"table", ev.Table,
+				"type", ev.EventType,
+				"gtid", ev.GTID)
+
+			batch = append(batch, ev)
+			if len(batch) >= idx.BatchSize() {
+				if err := flush(); err != nil {
+					return err
+				}
+			}
+		}
+	}
+}
+
+// ─── Config / One ───────────────────────────────────────────────
+
+// Config carries everything ONE replication stream needs — the by-value
+// equivalent of the strm* package globals. It exists so the control plane can
+// run N streams concurrently in one process (each with its own config), while
+// `bintrail stream` keeps wiring its flags through a single instance
+// unchanged. Plain data plus optional observer hooks and required Deps: the
+// zero value is invalid; build it from the cobra layer's streamConfigFromFlags
+// or populate it explicitly.
+type Config struct {
+	IndexDSN    string
+	SourceDSN   string
+	ServerID    uint32
+	StartFile   string
+	StartPos    uint32
+	StartGTID   string
+	BatchSize   int
+	Schemas     string
+	Tables      string
+	Checkpoint  int // seconds
+	MetricsAddr string
+	// MetricsSource is the value of the Prometheus "source" label for this
+	// stream — the supervisor sets it to the registry entry ID so N
+	// concurrent streams stay distinguishable on one /metrics endpoint.
+	// Empty = fall back to the resolved bintrail_id (or "default").
+	MetricsSource string
+	SSLMode       string
+	SSLCA         string
+	SSLCert       string
+	SSLKey        string
+	Format        string
+	Reset         bool
+	NoGapFill     bool
+	GapTimeout    int // seconds
+	// Hooks, when non-nil, lets a supervisor observe this stream's liveness
+	// without polling. Plain `bintrail stream` leaves it nil.
+	Hooks *Hooks
+	// Deps are the host-supplied helper functions One needs (see Deps). All
+	// fields are required.
+	Deps Deps
+}
+
+// Hooks are liveness callbacks a supervisor attaches to one stream.
+// All fields are optional. They are invoked synchronously from the stream
+// loop — implementations must be fast and non-blocking.
+type Hooks struct {
+	// OnCheckpoint fires after every successful checkpoint save — the
+	// "attached and healthy" signal (the ticker checkpoints even with zero
+	// events, so an idle source still reports progress).
+	OnCheckpoint func()
+	// OnIndexed fires after every successful batch flush with rows written.
+	OnIndexed func(n int64)
+	// OnGapAutoAdvance fires when an unfillable binlog gap forced the stream
+	// to advance past purged events — data in the gap is permanently lost.
+	OnGapAutoAdvance func(detail string)
+}
+
+// Deps are the host-supplied functions One needs that are NOT part of the
+// streaming engine itself — source preflight/validation, server-identity
+// resolution, schema snapshot/resolver/filters, DDL schema-change recording,
+// source-DSN parsing, and JSON summary output. They live in the binary that
+// builds the Config so streamrun stays a pure engine and never imports the
+// cmd layer (or anything it would drag in). Every field must be set.
+type Deps struct {
+	ValidateBinlogFormat   func(db *sql.DB) error
+	ValidateBinlogRowImage func(db *sql.DB) error
+	ValidateNoFKCascades   func(db *sql.DB, schemas []string) error
+	ParseSchemaList        func(s string) []string
+	ResolveServerIdentity  func(ctx context.Context, sourceDB, indexDB *sql.DB, sourceDSN string) (string, error)
+	EnsureResolver         func(indexDB, sourceDB *sql.DB, schemas []string) (*metadata.Resolver, error)
+	BuildIndexFilters      func(schemas, tables string) parser.Filters
+	InsertSchemaChange     func(db *sql.DB, ev parser.Event, snapshotID *int) error
+	ParseSourceDSN         func(dsn string) (host string, port uint16, user, password string, err error)
+	OutputJSON             func(v any) error
+}
+
+// One runs one complete replication stream — connect, validate,
+// resolve identity, snapshot, gap-check, sync, index, checkpoint — until ctx
+// is cancelled or a fatal error occurs. It is self-contained by design: no
+// package globals, no signal handling, safe to run N instances concurrently
+// (each against its own index database).
+func One(ctx context.Context, cfg Config) error {
+	if !cliutil.IsValidOutputFormat(cfg.Format) {
+		return fmt.Errorf("invalid --format %q; must be text or json", cfg.Format)
+	}
+	// Reject non-positive --gap-timeout values: a zero or negative timeout
+	// would produce an immediately-cancelled context inside detectPositionGap
+	// / detectGTIDGap, surfacing as a misleading "context deadline exceeded"
+	// error whose recovery hint (`--reset`) discards the saved checkpoint.
+	if cfg.GapTimeout <= 0 {
+		return fmt.Errorf("invalid --gap-timeout %d: must be a positive number of seconds", cfg.GapTimeout)
+	}
+
+	// Derived cancel: internal failures (e.g. the stream loop erroring) must
+	// stop the parser goroutine even when the caller's ctx stays live.
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	// ── 1. Connect to index database ─────────────────────────────────────────
+	indexDB, err := config.Connect(cfg.IndexDSN)
+	if err != nil {
+		return fmt.Errorf("failed to connect to index database: %w", err)
+	}
+	defer indexDB.Close()
+
+	if err := indexer.EnsureSchema(indexDB); err != nil {
+		return fmt.Errorf("schema migration: %w", err)
+	}
+
+	// ── 2. Connect to source database: validate binlog_row_image ─────────────
+	sourceDB, err := config.Connect(cfg.SourceDSN)
+	if err != nil {
+		return fmt.Errorf("failed to connect to source MySQL: %w", err)
+	}
+	defer sourceDB.Close()
+
+	if err := cfg.Deps.ValidateBinlogFormat(sourceDB); err != nil {
+		return err
+	}
+	fmt.Println("Source: binlog_format=ROW \u2713")
+
+	if err := cfg.Deps.ValidateBinlogRowImage(sourceDB); err != nil {
+		return err
+	}
+	fmt.Println("Source: binlog_row_image=FULL \u2713")
+
+	if err := cfg.Deps.ValidateNoFKCascades(sourceDB, cfg.Deps.ParseSchemaList(cfg.Schemas)); err != nil {
+		return err
+	}
+	fmt.Println("Source: no FK cascades \u2713")
+
+	// ── 3. Resolve server identity ────────────────────────────────────────────
+	bintrailID, err := cfg.Deps.ResolveServerIdentity(ctx, sourceDB, indexDB, cfg.SourceDSN)
+	if err != nil {
+		if errors.Is(err, serverid.ErrConflict) {
+			return fmt.Errorf("cannot stream: %w", err)
+		}
+		slog.Warn("server identity resolution failed; proceeding without bintrail_id", "error", err)
+	} else {
+		slog.Info("server identity resolved", "bintrail_id", bintrailID)
+	}
+
+	// ── 4. Schema snapshot + resolver ─────────────────────────────────────────
+	resolver, err := cfg.Deps.EnsureResolver(indexDB, sourceDB, cfg.Deps.ParseSchemaList(cfg.Schemas))
+	if err != nil {
+		return err
+	}
+	fmt.Printf("Snapshot: id=%d, tables=%d\n", resolver.SnapshotID(), resolver.TableCount())
+
+	// ── 5. Filters ────────────────────────────────────────────────────────────
+	filters := cfg.Deps.BuildIndexFilters(cfg.Schemas, cfg.Tables)
+
+	// ── 6. Determine start position ───────────────────────────────────────────
+	saved, err := loadStreamState(indexDB)
+	if err != nil {
+		return fmt.Errorf("failed to load stream state: %w", err)
+	}
+
+	if cfg.Reset {
+		if saved != nil {
+			if _, err := indexDB.Exec(`DELETE FROM stream_state WHERE id = 1`); err != nil {
+				return fmt.Errorf("failed to reset stream state: %w", err)
+			}
+			slog.Warn("cleared saved checkpoint (--reset)", "old_mode", saved.mode,
+				"old_file", saved.binlogFile, "old_pos", saved.binlogPos)
+			saved = nil
+		} else {
+			slog.Info("--reset specified but no saved checkpoint exists; ignoring")
+		}
+	}
+
+	mode, startFile, startGTIDStr, startPos, accGTID, err := resolveStartWithAutoDiscover(
+		cfg.StartFile, cfg.StartGTID, cfg.StartPos, saved,
+		func() (string, uint32, error) { return config.CurrentBinlogPosition(sourceDB) })
+	if err != nil {
+		return err
+	}
+	// Surface the auto-discovered position when this was a first-run, no-flags
+	// invocation. Mirrors the agent BYOS startup checkmark style.
+	if saved == nil && cfg.StartFile == "" && cfg.StartGTID == "" && mode == "position" {
+		slog.Info("auto-discovered current binlog position", "file", startFile, "pos", startPos)
+		fmt.Printf("Start position: auto-discovered %s:%d ✓\n", startFile, startPos)
+	}
+
+	// ── 6b. Detect binlog gap ────────────────────────────────────────────
+	// Only check for gaps when resuming from a saved checkpoint (not on first run).
+	if saved != nil {
+		var gap *gapResult
+		var gapErr error
+
+		gapTimeout := time.Duration(cfg.GapTimeout) * time.Second
+
+		switch mode {
+		case "position":
+			gap, gapErr = detectPositionGap(sourceDB, startFile, startPos, gapTimeout)
+		case "gtid":
+			gap, gapErr = detectGTIDGap(sourceDB, startGTIDStr, gapTimeout)
+		default:
+			slog.Warn("gap detection not implemented for mode", "mode", mode)
+		}
+
+		if gapErr != nil {
+			// Gap detection failure means we cannot verify whether binlogs have
+			// been purged. Fail hard — proceeding could mask data loss.
+			slog.Error("gap detection failed", "error", gapErr)
+			return fmt.Errorf("gap detection failed: %w (use --reset to skip gap detection and start from a new position)", gapErr)
+		} else if gap != nil && gap.HasGap {
+			if gap.Fillable {
+				slog.Info(gap.Message)
+				fmt.Println("Gap: fillable — replaying missed events before live tailing")
+			} else {
+				// Unfillable gap — binlogs/GTIDs have been purged.
+				slog.Warn(gap.Message)
+
+				if cfg.NoGapFill {
+					return fmt.Errorf("binlog gap detected and --no-gap-fill is set: %s", gap.Message)
+				}
+
+				// Auto-advance past the gap.
+				switch mode {
+				case "position":
+					slog.Warn("auto-advancing to earliest available binlog",
+						"old_file", startFile, "old_pos", startPos,
+						"new_file", gap.EarliestFile, "new_pos", gap.EarliestPos)
+					fmt.Printf("Gap: UNFILLABLE — advancing from %s:%d to %s:%d (events in between are permanently lost)\n",
+						startFile, startPos, gap.EarliestFile, gap.EarliestPos)
+					startFile = gap.EarliestFile
+					startPos = gap.EarliestPos
+
+				case "gtid":
+					slog.Warn("auto-advancing past purged GTIDs",
+						"old_gtid_set", startGTIDStr,
+						"purged_gtid_set", gap.PurgedGTIDSet)
+					fmt.Printf("Gap: UNFILLABLE — checkpoint GTID set includes purged GTIDs; advancing past purged set (events are permanently lost)\n")
+					// Use the purged set as the checkpoint GTID set — this tells
+					// MySQL we have already seen all purged GTIDs, so it will
+					// only send the non-purged GTIDs that remain in @@gtid_executed.
+					startGTIDStr = NormalizeGTIDSet(gap.PurgedGTIDSet)
+					gs, parseErr := gomysql.ParseMysqlGTIDSet(startGTIDStr)
+					if parseErr != nil {
+						return fmt.Errorf("failed to parse purged GTID set for auto-advance: %w", parseErr)
+					}
+					parsed, ok := gs.(*gomysql.MysqlGTIDSet)
+					if !ok {
+						return fmt.Errorf("unexpected GTID set type %T after parsing purged set", gs)
+					}
+					accGTID = parsed
+				}
+
+				// Persist the advanced position immediately so that if the stream
+				// crashes during startup, the next restart won't hit the same
+				// purged-binlog error again.
+				advancedState := &streamState{
+					mode:          mode,
+					binlogFile:    startFile,
+					binlogPos:     uint64(startPos),
+					gtidSet:       startGTIDStr,
+					serverID:      cfg.ServerID,
+					bintrailID:    bintrailID,
+					eventsIndexed: saved.eventsIndexed,
+					lastEventTime: saved.lastEventTime,
+				}
+				if err := saveCheckpoint(indexDB, advancedState); err != nil {
+					return fmt.Errorf("failed to save advanced checkpoint: %w", err)
+				}
+				slog.Info("saved advanced checkpoint after gap auto-advance",
+					"file", startFile, "pos", startPos, "gtid_set", startGTIDStr)
+				// Record the data loss DURABLY: the advanced checkpoint means
+				// the next start sees no gap, so without this row the only
+				// trace of the permanently lost events would be an in-memory
+				// flag (or a log line) that a restart silently discards.
+				// Cleared by an explicit monitor Stop or --reset.
+				if _, err := indexDB.Exec(`UPDATE stream_state
+					SET gap_lost_at = UTC_TIMESTAMP(), gap_lost_detail = ?
+					WHERE id = 1`, gap.Message); err != nil {
+					slog.Warn("failed to persist gap-loss record; the lost_position state will not survive a restart", "error", err)
+				}
+				// Tell the supervisor events were lost — without this the
+				// auto-advance is only a log line and the console shows a
+				// healthy RUNNING badge over a stream that silently skipped
+				// data (#402).
+				if cfg.Hooks != nil && cfg.Hooks.OnGapAutoAdvance != nil {
+					cfg.Hooks.OnGapAutoAdvance(gap.Message)
+				}
+			}
+		}
+	}
+
+	state := &streamState{
+		mode: mode,
+		// Seed the position with the resolved start so the first ticker
+		// checkpoint (which fires even before any event arrives) persists a
+		// valid resume point instead of an empty file / position 0.
+		binlogFile: startFile,
+		binlogPos:  uint64(startPos),
+		serverID:   cfg.ServerID,
+		accGTID:    accGTID,
+		bintrailID: bintrailID,
+	}
+	if saved != nil {
+		state.eventsIndexed = saved.eventsIndexed
+	}
+	if startGTIDStr != "" {
+		state.gtidSet = startGTIDStr
+	}
+
+	// ── 7. Parse source DSN for BinlogSyncer ─────────────────────────────
+	host, port, user, password, err := cfg.Deps.ParseSourceDSN(cfg.SourceDSN)
+	if err != nil {
+		return err
+	}
+
+	// ── 6b. Build TLS config ──────────────────────────────────────────────────
+	tlsCfg, err := buildTLSConfig(cfg.SSLMode, cfg.SSLCA, cfg.SSLCert, cfg.SSLKey, host)
+	if err != nil {
+		return err
+	}
+
+	// ── 7. Create BinlogSyncer ────────────────────────────────────────────────────
+	syncerCfg := replication.BinlogSyncerConfig{
+		ServerID:             cfg.ServerID,
+		Flavor:               "mysql",
+		Host:                 host,
+		Port:                 port,
+		User:                 user,
+		Password:             password,
+		HeartbeatPeriod:      30 * time.Second,
+		MaxReconnectAttempts: 0, // infinite retry
+		TLSConfig:            tlsCfg,
+	}
+
+	// Use a closure defer so the active syncer is always closed on exit,
+	// even if we replace it during the preferred-mode TLS fallback below.
+	// The nil guard prevents a panic if an early-return is added before
+	// syncer is assigned.
+	var syncer *replication.BinlogSyncer
+	defer func() {
+		if syncer != nil {
+			syncer.Close()
+		}
+	}()
+	syncer = replication.NewBinlogSyncer(syncerCfg)
+
+	// startStreamer starts sync from the resolved position/GTID set.
+	startStreamer := func() (*replication.BinlogStreamer, error) {
+		switch mode {
+		case "position":
+			s, startErr := syncer.StartSync(gomysql.Position{Name: startFile, Pos: startPos})
+			if startErr != nil {
+				return nil, fmt.Errorf("StartSync(%s, %d): %w", startFile, startPos, startErr)
+			}
+			return s, nil
+		case "gtid":
+			gset, parseErr := gomysql.ParseGTIDSet("mysql", startGTIDStr)
+			if parseErr != nil {
+				return nil, fmt.Errorf("parse start GTID set: %w", parseErr)
+			}
+			s, startErr := syncer.StartSyncGTID(gset)
+			if startErr != nil {
+				return nil, fmt.Errorf("StartSyncGTID: %w", startErr)
+			}
+			return s, nil
+		default:
+			return nil, fmt.Errorf("unexpected mode %q", mode)
+		}
+	}
+
+	// ── 8. Start sync ───────────────────────────────────────────────────────────────
+	streamer, startErr := startStreamer()
+	if startErr != nil && cfg.SSLMode == "preferred" {
+		// preferred: TLS attempt failed — retry without TLS.
+		slog.Warn("initial connection failed; retrying without TLS (--ssl-mode preferred)", "error", startErr)
+		syncer.Close()
+		syncerCfg.TLSConfig = nil
+		syncer = replication.NewBinlogSyncer(syncerCfg)
+		streamer, startErr = startStreamer()
+		if startErr != nil {
+			return startErr
+		}
+	} else if startErr != nil {
+		return startErr
+	}
+
+	switch mode {
+	case "position":
+		fmt.Printf("Streaming from %s position %d\n", startFile, startPos)
+	case "gtid":
+		fmt.Printf("Streaming from GTID set: %s\n", startGTIDStr)
+	}
+
+	// (Signal handling lives in runStream — the process owner. One only
+	// honors ctx, so a supervisor can run several instances under one
+	// lifecycle without competing signal handlers.)
+
+	// ── 9. Optional Prometheus metrics HTTP server ─────────────────────────
+	// Under `up --console` the daemon serves one endpoint for all streams
+	// instead (cfg.MetricsAddr is empty there) — the registry is process-
+	// global, so one handler exposes every per-source series.
+	if cfg.MetricsAddr != "" {
+		stopMetrics, err := StartMetricsServer(cfg.MetricsAddr)
+		if err != nil {
+			return err
+		}
+		defer stopMetrics()
+	}
+
+	// All stream metrics carry a "source" label: the supervisor's entry ID
+	// when monitored, else the resolved bintrail_id ("default" if unknown).
+	metricsSource := cfg.MetricsSource
+	if metricsSource == "" {
+		metricsSource = bintrailID
+	}
+	metrics := observe.ForSource(metricsSource)
+
+	// ── 10. StreamParser + its synchronous DDL hook ──────────────────────────
+	sp := parser.NewStreamParser(resolver, filters, nil)
+	idx := indexer.New(indexDB, cfg.BatchSize)
+
+	// ── 11. DDL auto-snapshot hook — registered BEFORE Run starts so even a
+	// DDL arriving in the first events cannot miss it ─────────────────────────
+	// Registered on the parser, not on streamLoop: the binlog is sequential
+	// (`CREATE TABLE t; INSERT INTO t;`), so the resolver must be refreshed
+	// before the parser decodes the events that FOLLOW the DDL — a
+	// consumer-side handler ran too late and the trailing rows were skipped
+	// as "table not in snapshot" (#396). Best-effort: failures are logged and
+	// streaming continues with the previous schema.
+	// TRUNCATE does not change schema structure, so it only records the change.
+	schemas := cfg.Deps.ParseSchemaList(cfg.Schemas)
+	sp.SetSyncDDLHook(func(ev parser.Event) {
+		if ev.DDLType == parser.DDLTruncateTable {
+			slog.Info("DDL detected (no snapshot needed)",
+				"file", ev.BinlogFile, "pos", ev.EndPos,
+				"ddl_type", ev.DDLType, "schema", ev.Schema, "table", ev.Table)
+			if err := cfg.Deps.InsertSchemaChange(indexDB, ev, nil); err != nil {
+				slog.Warn("failed to record schema change", "error", err)
+			}
+			return
+		}
+
+		slog.Info("DDL detected — taking auto-snapshot",
+			"file", ev.BinlogFile, "pos", ev.EndPos,
+			"ddl_type", ev.DDLType, "schema", ev.Schema, "table", ev.Table)
+
+		stats, snapErr := metadata.TakeSnapshot(sourceDB, indexDB, schemas)
+		var snapID *int
+		if snapErr != nil {
+			slog.Error("auto-snapshot after DDL failed; subsequent events may use stale schema",
+				"error", snapErr, "ddl_type", ev.DDLType, "table", ev.Table)
+		} else {
+			snapID = &stats.SnapshotID
+			newResolver, resolverErr := metadata.NewResolver(indexDB, stats.SnapshotID)
+			if resolverErr != nil {
+				slog.Warn("failed to load new resolver after DDL snapshot", "error", resolverErr)
+			} else {
+				sp.SwapResolver(newResolver)
+				slog.Info("auto-snapshot taken; resolver updated",
+					"snapshot_id", stats.SnapshotID,
+					"tables", stats.TableCount,
+					"columns", stats.ColumnCount)
+			}
+		}
+
+		if err := cfg.Deps.InsertSchemaChange(indexDB, ev, snapID); err != nil {
+			slog.Warn("failed to record schema change", "error", err)
+		}
+	})
+
+	events := make(chan parser.Event, 1000)
+	parseErrCh := make(chan error, 1)
+
+	go func() {
+		defer close(events)
+		parseErrCh <- sp.Run(ctx, streamer, events)
+	}()
+
+	// ── 12. Run stream loop with checkpointing ──────────────────────────────────
+	fmt.Printf("Streaming started (server-id=%d, checkpoint=%ds)\n", cfg.ServerID, cfg.Checkpoint)
+	loopErr := streamLoop(ctx, events, idx, indexDB,
+		time.Duration(cfg.Checkpoint)*time.Second, state, metrics, cfg.Hooks)
+
+	parseErr := <-parseErrCh
+
+	// ── 12. Summary ───────────────────────────────────────────────────────────────
+	if loopErr != nil {
+		return loopErr
+	}
+	if parseErr != nil && !errors.Is(parseErr, context.Canceled) {
+		return parseErr
+	}
+
+	if cfg.Format == "json" {
+		return cfg.Deps.OutputJSON(struct {
+			EventsIndexed int64  `json:"events_indexed"`
+			LastFile      string `json:"last_file"`
+			LastPosition  uint64 `json:"last_position"`
+		}{
+			EventsIndexed: state.eventsIndexed,
+			LastFile:      state.binlogFile,
+			LastPosition:  state.binlogPos,
+		})
+	}
+
+	fmt.Printf("\nEvents indexed: %d\n", state.eventsIndexed)
+	fmt.Printf("Last position:  %s:%d\n", state.binlogFile, state.binlogPos)
+	return nil
+}
+
+// StartMetricsServer binds the Prometheus /metrics endpoint and returns a
+// shutdown func. The bind is SYNCHRONOUS so a bad --metrics-addr fails the
+// command fast — the operator explicitly asked for metrics; silently running
+// without them (scrapes getting connection-refused) is worse than refusing
+// to start. Shared by `bintrail stream` (one endpoint per stream process)
+// and the `up --console` daemon (one endpoint for all supervised streams;
+// the default registry is process-global).
+func StartMetricsServer(addr string) (shutdown func(), err error) {
+	mux := http.NewServeMux()
+	mux.Handle("/metrics", promhttp.Handler())
+	srv := &http.Server{Handler: mux}
+	ln, err := net.Listen("tcp", addr)
+	if err != nil {
+		return nil, fmt.Errorf("metrics server: cannot bind %s: %w", addr, err)
+	}
+	go func() {
+		slog.Info("metrics server starting", "addr", addr)
+		if serveErr := srv.Serve(ln); serveErr != nil && !errors.Is(serveErr, http.ErrServerClosed) {
+			slog.Error("metrics server error", "error", serveErr)
+		}
+	}()
+	return func() {
+		shutCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		_ = srv.Shutdown(shutCtx)
+	}, nil
+}
